updates to native packages; consolidate ts files for tsconfig src to include fluvio.spec.ts and move to src file
