+++ native/src/darwin/package.json
@@ -3,6 +3,9 @@
   "version": "0.1.0-alpha.1",
   "description": "Pre-built MacOS Fluvio client for Node.js",
   "main": "./index.node",
+  "files": [
+    "./index.node"
+  ],
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
