+++ package.json
@@ -9,7 +9,7 @@
         "client"
     ],
     "files": [
-        "./src/**/*.ts"
+        "./src/index.ts"
     ],
     "main": "./dist/index.js",
     "types": "./dist/index.d.ts",
@@ -20,12 +20,14 @@
     "homepage": "https://fluvio.io",
     "author": "team@fluvio.io",
     "license": "Apache-2.0",
-    "dependencies": {},
+    "dependencies": {
+        "@fluvio/native": "file:native"
+    },
     "scripts": {
-        "prettier": "npx prettier --write '{src,examples,test,demos}/**/*.{ts,js}'",
-        "prettier:check": "npx prettier --check '{src,examples,test,demos}/**/*.{ts,js}'",
+        "prettier": "npx prettier --write '{src,examples,test,native,demos}/**/*.{ts,js}'",
+        "prettier:check": "npx prettier --check '{src,examples,test,native,demos}/**/*.{ts,js}'",
         "lint": "npx tslint -c tslint.json '{src,examples,test}/**/*.ts'",
-        "tsc": "npx tsc ./src/index.ts --outDir ./dist -d",
+        "tsc": "npx tsc -p .",
         "build": "rm -rf ./dist/ && npm run tsc",
         "build:darwin": "nj-cli build -o ./native/src/darwin",
         "build:linux": "nj-cli build -o ./native/src/linux",
