+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
   "exclude": [
     "./examples/**/*",
-    "./test/**/*",
     "./demos/**/*",
     "./dist/**/*",
     "./build.js",
@@ -9,7 +8,7 @@
     "./docs/**/*",
   ],
   "compilerOptions": {
-    "target": "ES2015",
+    "target": "es6",
     "module": "commonjs",
     "allowJs": true,
     "declaration": true,
